+++ Sources/KeyboardKit/Features/FeatureToggle.swift
@@ -14,11 +14,6 @@ import Foundation
 
  You can use the static ``shared`` context to share a single
  instance in your code.
-
- Use ``toggleFeature(_:_:)`` to toggle any feature on or off
- and ``isFeatureEnabled(_:)`` to see if a feature is enabled
- or not. ``enabledFeatures`` returns a list with all enabled
- features and ``Feature/allFeatures`` all available features.
  */
 public final class FeatureToggle {